requirements.txt
django
djangorestframework
dj-database-url
gunicorn
markdown
whitenoise
